refactor: move all annotation-related logic to `annotation_builder/`
