+++ lua/codedocs/annotation_builder/builder_spec.lua
@@ -36,7 +36,7 @@ describe("Annotation builder - ", function()
 			local expected_output = vim.fn.readfile(DIR .. "/cases/" .. case_name .. "/output")
 
 			local opts = require("codedocs.config").opts.annot_builder
-			local lines = Codedocs.build_annot_lines(blocks, opts, 1, MOCKED_ITEMS)
+			local lines = require("codedocs.annotation_builder").build_annot_lines(blocks, opts, 1, MOCKED_ITEMS)
 
 			assert.are.same(expected_output, lines)
 		end)
