+++ lua/codedocs/init.lua
@@ -3,13 +3,6 @@ local Item_extractor = require "codedocs.item_extractor"
 
 local Codedocs = {}
 
-function Codedocs.build_annot_lines(blocks, opts, row, items)
-	local annot = require("codedocs.annotation_builder").new(items, opts, row)
-	local lines = annot:build(blocks)
-
-	return lines
-end
-
 --- Inserts an annotation relative to a target and moves the cursor to the annotation title
 ---@param annotation_lines string[]
 ---@param row number 0-based annotation-related positions
@@ -41,47 +34,7 @@ local function _write_to_buffer(annotation_lines, row, placement)
 	vim.snippet.expand(lines)
 end
 
-local function _determine_lang_name()
-	if not Codedocs._filetypes_map then
-		local langs_config = require("codedocs.config").opts.languages
-		local filetypes_map = {}
-		for lang_name, opts in pairs(langs_config) do
-			for _, filetype_name in ipairs(opts.filetypes) do
-				filetypes_map[filetype_name] = lang_name
-			end
-		end
-
-		Codedocs._filetypes_map = filetypes_map
-	end
-
-	return Codedocs._filetypes_map[vim.bo.filetype]
-end
-
 function Codedocs.get_supported_langs() return vim.tbl_keys(require("codedocs.config").opts.languages) end
-
-function Codedocs.get_annot_list()
-	local lang = _determine_lang_name()
-	local lang_stuff = require("codedocs.config").opts.languages[lang]
-	local annot_names = vim.tbl_keys(lang_stuff.styles[lang_stuff.default_style].annots)
-	return annot_names
-end
-
----Returns an existing annotation table from the configuration table
----Equivalent to `require("codedocs.config").opts.languages[[lang_name]].styles[[style_name]].annots[[annotation_name]]
----@param lang_name string
----@param style_name string
----@param annot_name string
----@return CodedocsAnnotationStyleOpts annotation_tbl
-function Codedocs.get_annot_tbl(lang_name, style_name, annot_name)
-	vim.validate {
-		lang_name = { lang_name, "string" },
-		style_name = { style_name, "string" },
-		annot_name = { annot_name, "string" },
-	}
-
-	local annot_tbl = require("codedocs.config").opts.languages[lang_name].styles[style_name].annots[annot_name]
-	return annot_tbl
-end
 
 ---@return string[] supported_styles List of style names
 function Codedocs.get_supported_styles(lang_name)
@@ -116,31 +69,6 @@ function Codedocs.get_target_data(lang_name, annotation_data)
 	return Item_extractor.get_detected_target_data(lang_name) or {}
 end
 
-function Codedocs.prepare_annotation(lang_name, annot_data)
-	local target_data = Codedocs.get_target_data(lang_name, annot_data)
-
-	local lang_config = require("codedocs.config").opts.languages[lang_name]
-	local style_name = annot_data and annot_data.style_name or lang_config.default_style
-
-	local annot_tbl = Codedocs.get_annot_tbl(lang_name, style_name, target_data.target_name)
-
-	local lines = Codedocs.build_annot_lines(
-		annot_tbl.blocks,
-		require("codedocs.config").opts.annot_builder,
-		target_data.row,
-		target_data.items
-	)
-
-	return {
-		row = target_data.row,
-		target_name = target_data.target_name,
-		style_name = style_name,
-		items = target_data.items,
-		placement = annot_tbl.placement,
-		lines = lines,
-	}
-end
-
 ---@param annot_data { annot_name: string, style_name: string? }?
 function Codedocs.generate(annot_data)
 	Logger.info "Annotation generation started"
@@ -158,10 +86,11 @@ function Codedocs.generate(annot_data)
 		Logger.info("Passed data: " .. vim.inspect(annot_data))
 	end
 
-	local lang_name = _determine_lang_name()
+	local lang_name = require("codedocs.utils.general")._determine_lang_name()
 	Logger.info("Language: " .. lang_name)
 
-	local annot = Codedocs.prepare_annotation(lang_name, annot_data)
+	local target_data = Codedocs.get_target_data(lang_name, annot_data)
+	local annot = require("codedocs.annotation_builder").prepare_annotation(lang_name, annot_data, target_data)
 
 	Logger.info("Annotation content" .. vim.inspect(annot.lines))
